refactor(core): move parse_utc_seconds/days_from_civil into time.rs (sc-8897)
time.rs owned civil_from_days (days -> Y/M/D) and format_unix_seconds, but
their inverses parse_utc_seconds + days_from_civil (Y/M/D -> days) lived
~8k lines away in jobs_store.rs. Relocate both beside format_unix_seconds
so each direction of the date conversion sits with its partner;
jobs_store now imports parse_utc_seconds from crate::time.

Kept the tolerated `.digitsZ` fractional-seconds suffix rather than
removing it: although neither this module's utc_now() nor the frozen
Python predecessor's (microsecond=0) ever emits sub-second digits, the
branch is a deliberately-tested capability
(elapsed_seconds_accepts_fractional_rfc3339_timestamps writes a fractional
timestamp straight into the DB and asserts it parses), so dropping it
would delete tested behavior, not just dead code. Documented the rationale
on the function. Added time.rs unit tests: format->parse round-trip
(epoch, leap-day, pre-epoch), malformed rejection, and fractional-second
truncation.
